feat: extract client-servers.ts, restore WebSocketBroadcastChannel constructor
